fyp requirement.txt
numpy==1.19.5
Pillow==8.1.0
tensorflow==2.4.0
keras==2.4.3
opencv==4.5.1
matplotlib==3.3.3